Add files via upload
How did COVID-19 affect inflation levels in Canada?

Overview 

The COVID-19 lockdowns implemented by the Federal government of Canada had an acute, direct, and immediate impact. 

The effects of COVID-19 lockdowns in 2020 and 2021 had resulted in a significant decrease in economic activity. More specifically, a reduction in the transportation of goods; and the limitation of access to or the suspension of services deemed non essential. 

To further dampen any adverse economic effects of the lockdowns, the federal government had also put a substantial amount of money into the hands of consumers, private businesses, and commercial banks. This also happened to be during a period of record low interest rates, set by the Bank of Canada.

This analysis will not look at the details of the supply chains. It will be assumed to be a given external variable that is, for the most part, out of the government's control. The labour market will not play a role in this analysis as well. The assumption is that the government had to interrupt, or nearly completely halt public activity during the public health crisis. And it is assumed that the labour market recovered after the lockdowns were declared to be over (This fact can be attested to by near record low unemployment numbers as of August 2022). And finally government sales tax rates, which are included in final sales prices, will be assumed to be steady throughout September 2019 and August 2022






This analysis will focus on inflation and other possible contributors to inflation rates from September of 2019 to August of 2022. Specifically, the private household debt levels of Canadians, and the interest rates set by the central bank of Canada will be explored for any effects on and correlations with inflation.


This analysis seeks to determine the following:-


How was inflation before, during and after the COVID-19 lockdowns? How much did Canadian household debt change through out this period? How much did interest rates change? 

       Inflation throughout Canada during this period.
       Household debt levels during this period.
       Interest rates through out this period.
       Demonstrate if there are any correlations between: 
       -Borrowing/debt and interest rates
       -Inflation and interest rates and debt levels 




This analysis demonstrates that:-

After the lockdowns were lifted, employment numbers improved rapidly. Due to continuing and ongoing international events, which are outside of the federal government's scope of control, the frequency of transportation and the quantities of the supplies of goods is currently insufficient. This has created a situation where there is an inadequate amount of a supply of goods to meet increased spending. 

A healthy labour market, high savings, high amounts of private borrowing, and a hamstrung supply chain have created historic levels of price inflation. There is more cash in the hands of Canadians today than there are goods and services available to meet demand. 

As a result, there is an upward and steep trend in the consumer price index. This is especially more evident during and after the gradual easing and ending of the COVID-19 lockdowns leading up to January of 2022 in most provinces and territories of Canada. 

It can also be seen that there is a substantial increase in the levels of different types of debts taken on by Canadians (i.e. mortgage and non-mortgage debts), during this period of historically low interest rates and high liquidity in circulation. 

The Bank of Canada has responded with record increases in interest rates, in an attempt to discourage spending and borrowing, to reign inflation in. The bank has a target inflation rate below 3%. The rate of inflation is well above this number. This analysis ends before the inflation problem is resolved by the Bank of Canada, so further increases in interest rates are highly likely after August 2022 (after the period of this study covers). 

In conclusion, it is abundantly evident that low bank rates, increased consumer spending have directly contributed to the increase in the rate of inflation in Canada.